web/requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
babel==2.8.0              # via flask-babelex
bcrypt==3.2.0             # via -r requirements.in
blinker==1.4              # via flask-mail, flask-principal
cffi==1.14.3              # via bcrypt
click==7.1.2              # via flask
dnspython==2.0.0          # via email-validator
email-validator==1.1.1    # via flask-security-too
flask-babelex==0.9.4      # via flask-security-too
flask-login==0.5.0        # via flask-security-too
flask-mail==0.9.1         # via flask-security-too
flask-principal==0.4.0    # via flask-security-too
flask-restless==0.17.0    # via -r requirements.in
flask-security-too==3.4.4  # via -r requirements.in
flask-sqlalchemy==2.4.4   # via -r requirements.in
flask-wtf==0.14.3         # via flask-security-too
flask==1.1.2              # via flask-babelex, flask-login, flask-mail, flask-principal, flask-restless, flask-security-too, flask-sqlalchemy, flask-wtf
gunicorn==20.0.4          # via -r requirements.in
idna==2.10                # via email-validator
itsdangerous==1.1.0       # via flask, flask-security-too, flask-wtf
jinja2==2.11.2            # via flask, flask-babelex
markupsafe==1.1.1         # via jinja2, wtforms
mimerender==0.6.0         # via flask-restless
passlib==1.7.4            # via flask-security-too
pycparser==2.20           # via cffi
python-dateutil==2.8.1    # via flask-restless
python-mimeparse==1.6.0   # via mimerender
pytz==2020.1              # via babel
six==1.15.0               # via bcrypt, python-dateutil
speaklater==1.3           # via flask-babelex
sqlalchemy==1.3.20        # via flask-restless, flask-sqlalchemy
werkzeug==1.0.1           # via flask
wtforms==2.3.3            # via flask-wtf

# The following packages are considered to be unsafe in a requirements file:
# setuptools
